Move Equalizer to a separated packate
