+++ app/src/main/java/com/yurii/youtubemusic/screens/saved/SavedMusicFragment.kt
@@ -7,7 +7,7 @@ import androidx.fragment.app.viewModels
 import androidx.lifecycle.Observer
 import com.google.android.material.tabs.TabLayoutMediator
 import com.yurii.youtubemusic.screens.categories.CategoriesEditorActivity
-import com.yurii.youtubemusic.EqualizerActivity
+import com.yurii.youtubemusic.screens.equalizer.EqualizerActivity
 import com.yurii.youtubemusic.R
 import com.yurii.youtubemusic.databinding.FragmentSavedMusicBinding
 import com.yurii.youtubemusic.models.Category
